split prediction pipeline
